requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
statsmodels>=0.13.0
scipy>=1.7.0
tqdm>=4.62.0

# Data handling
pyarrow>=7.0.0
pandas-ta>=0.3.0b0
yfinance>=0.1.70
nsepy>=0.8
nsetools>=1.0.11
pyalgotrade>=0.20
ta>=0.9.0

# Web and API
flask>=2.0.0
flask-restful>=0.3.9
flask-cors>=3.0.10
dash>=2.5.0
dash-bootstrap-components>=1.1.0
plotly>=5.8.0
requests>=2.27.0
websocket-client>=1.3.2

# Broker APIs
kiteconnect>=4.1.0

# Database
sqlalchemy>=1.4.0
pymongo>=4.1.0
influxdb-client>=1.26.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Utilities
python-dateutil>=2.8.2
joblib>=1.1.0
python-dotenv>=0.20.0
loguru>=0.6.0
rich>=12.0.0
